@@ -5,6 +5,7 @@
   "license": "MIT",
   "repository": "GMOD/indexedfasta-js",
   "main": "dist/index.js",
+  "module": "esm/index.js",
   "author": {
     "name": "Colin Diesh",
     "email": "colin.diesh@gmail.com",
@@ -17,22 +18,22 @@
     "lib": "dist"
   },
   "files": [
-    "dist"
+    "dist",
+    "esm"
   ],
   "scripts": {
     "test": "jest",
     "coverage": "npm test -- --coverage",
     "lint": "eslint src test",
     "docs": "documentation readme src --section=API",
-    "clean": "rimraf dist",
+    "clean": "rimraf dist esm",
     "prebuild": "npm run clean",
-    "build": "babel src -d dist",
+    "build:esm": "tsc --target es2018 --outDir esm",
+    "build:es5": "tsc --target es5 --outDir dist",
+    "build": "npm run build:esm && npm run build:es5",
     "version": "standard-changelog && git add CHANGELOG.md",
     "prepublishOnly": "npm test && npm run build",
     "postversion": "git push origin master --follow-tags"
-  },
-  "jest": {
-    "testEnvironment": "node"
   },
   "keywords": [
     "fasta",
@@ -41,34 +42,37 @@
     "bionode"
   ],
   "dependencies": {
-    "@babel/runtime": "^7.12.5",
-    "@gmod/bgzf-filehandle": "^1.2.4",
-    "browser-or-node": "^1.3.0",
+    "@gmod/bgzf-filehandle": "^1.4.0",
     "es6-promisify": "^6.0.1",
     "object.fromentries": "^2.0.4"
   },
   "devDependencies": {
-    "@babel/cli": "^7.4.3",
-    "@babel/core": "^7.4.3",
-    "@babel/plugin-transform-runtime": "^7.4.3",
-    "@babel/preset-env": "^7.4.3",
-    "babel-eslint": "^10.0.1",
-    "babel-jest": "^24.7.0",
+    "@types/es6-promisify": "^6.0.0",
+    "@types/jest": "^27.0.3",
+    "@types/object.fromentries": "^2.0.1",
+    "@typescript-eslint/eslint-plugin": "^5.7.0",
+    "@typescript-eslint/parser": "^5.7.0",
     "cross-fetch": "^3.0.0",
     "documentation": "^9.1.1",
-    "eslint": "^5.12.0",
-    "eslint-config-airbnb-base": "^13.1.0",
-    "eslint-config-prettier": "^4.1.0",
-    "eslint-plugin-import": "^2.17.2",
-    "eslint-plugin-prettier": "^3.0.1",
-    "jest": "^24.7.0",
-    "jest-cli": "^24.7.0",
-    "prettier": "^1.17.0",
+    "eslint": "^7.0.0",
+    "eslint-config-airbnb-base": "^15.0.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-import": "^2.25.3",
+    "eslint-plugin-prettier": "^4.0.0",
+    "generic-filehandle": "^2.2.1",
+    "jest": "^27.4.5",
+    "prettier": "^2.5.1",
     "quick-lru": "^2.0.0",
     "rimraf": "^2.6.1",
-    "standard-changelog": "^1.0.0"
+    "standard-changelog": "^1.0.0",
+    "ts-jest": "^27.1.1",
+    "typescript": "^4.5.4"
   },
   "publishConfig": {
     "access": "public"
+  },
+  "browser": {
+    "./dist/localFile.js": false,
+    "./esm/localFile.js": false
   }
 }
