@@ -1,4 +1,5 @@
 import fromEntries from 'object.fromentries'
+import type { GenericFilehandle } from 'generic-filehandle'
 import LocalFile from './localFile'
 import BgzipIndexedFasta from './bgzipIndexedFasta'
 import IndexedFasta from './indexedFasta'
@@ -8,7 +9,7 @@ if (!Object.fromEntries) {
   fromEntries.shim()
 }
 
-function parseSmallFasta(text) {
+function parseSmallFasta(text: string) {
   return text
     .split('>')
     .filter(t => /\S/.test(t))
@@ -26,11 +27,17 @@ function parseSmallFasta(text) {
 
 // memoized
 class FetchableSmallFasta {
-  constructor({ fasta, path }) {
+  fasta: GenericFilehandle
+
+  data: Promise<{ id: string; description: string; sequence: string }[]>
+
+  constructor({ fasta, path }: { fasta: GenericFilehandle; path: string }) {
     if (fasta) {
       this.fasta = fasta
     } else if (path) {
       this.fasta = new LocalFile(path)
+    } else {
+      throw new Error('Need to pass fasta or path')
     }
     this.data = this.fasta.readFile().then(buffer => {
       const text = buffer.toString('utf8')
@@ -38,11 +45,13 @@ class FetchableSmallFasta {
     })
   }
 
-  async fetch(id, start, end) {
+  async fetch(id: string, start: number, end: number) {
     const data = await this.data
     const entry = data.find(iter => iter.id === id)
     const length = end - start
-    if (!entry) throw new Error(`no sequence with id ${id} exists`)
+    if (!entry) {
+      throw new Error(`no sequence with id ${id} exists`)
+    }
     return entry.sequence.substr(start, length)
   }
 
